problem after receiving a number of chars

+++ utils.c
@@ -22,7 +22,7 @@ void	ft_putchar(char c)
 	write(1, &c, 1);
 }
 
-void	ft_putnbr(size_t n)
+void	ft_putnbr(long int n)
 {
 	if (n < 0)
 	{
