+++ utils.h
@@ -10,7 +10,7 @@ int		ft_atoi(const char *s);
 size_t	ft_strlen(char *s);
 void	ft_putstr(char *s);
 void	ft_putchar(char c);
-void	ft_putnbr(size_t n);
+void	ft_putnbr(long int n);
 void	signal_handler(int sign_int);
 
 #endif
